+++ src/Database/Reader.php
@@ -110,6 +110,27 @@ class Reader implements ProviderInterface
     }
 
     /**
+     * This method returns a GeoLite2 ASN model.
+     *
+     * @param string $ipAddress IPv4 or IPv6 address as a string.
+     *
+     * @return \GeoIp2\Model\Asn
+     *
+     * @throws \GeoIp2\Exception\AddressNotFoundException if the address is
+     *         not in the database.
+     * @throws \MaxMind\Db\Reader\InvalidDatabaseException if the database
+     *         is corrupt or invalid
+     */
+    public function asn($ipAddress)
+    {
+        return $this->flatModelFor(
+            'Asn',
+            'GeoLite2-ASN',
+            $ipAddress
+        );
+    }
+
+    /**
      * This method returns a GeoIP2 Connection Type model.
      *
      * @param string $ipAddress IPv4 or IPv6 address as a string.